// frontend/src/api/generated/newChatbotAPI.schemas.ts
export type GetAdminProducts200AllOfTwoDataItemAllOf = {
  id: string;
  /**
   * @minLength 2
   * @maxLength 100
   */
  name: string;
  description?: string;
  features: string[];
  active: boolean;
  archived: boolean;
  createdAt: string;
  updatedAt: string;
};
// ...
export type GetAdminProducts200AllOfTwoDataItemAllOfTwoPricesItem = {
  id: string;
  productId: string;
  amount: number;
  currency: typeof GetAdminProducts200AllOfTwoDataItemAllOfTwoPricesItemCurrency[keyof typeof GetAdminProducts200AllOfTwoDataItemAllOfTwoPricesItemCurrency] ;
  stripePriceId?: string;
  active: boolean;
  archived: boolean;
  intervalType: typeof GetAdminProducts200AllOfTwoDataItemAllOfTwoPricesItemIntervalType[keyof typeof GetAdminProducts200AllOfTwoDataItemAllOfTwoPricesItemIntervalType] ;
  /**
   * @minimum 1
   * @maximum 1000
   */
  intervalValue: number;
  createdAt: string;
  updatedAt: string;
};

export type GetAdminProducts200AllOfTwoDataItemAllOfTwo = {
  prices?: GetAdminProducts200AllOfTwoDataItemAllOfTwoPricesItem[];
};

export type GetAdminProducts200AllOfTwoDataItem = GetAdminProducts200AllOfTwoDataItemAllOf & GetAdminProducts200AllOfTwoDataItemAllOfTwo;